fix build warning: comparison between signed and unsigned integer

+++ test/TestUtils.cc
@@ -49,7 +49,7 @@ TEST(Utils, splitNotify) {
 
   const char *pch = splitNotify(line);
 
-  ASSERT_EQ(pch - line.c_str(), l1.length());
+  ASSERT_EQ((size_t)(pch - line.c_str()), l1.length());
   ASSERT_EQ(strlen(pch), l2.length());
 }
 
@@ -62,9 +62,9 @@ TEST(Utils, Strings_parseConfJson) {
     ASSERT_EQ(parseConfJson(line, listenIP, listenPort, poolConfs), true);
     ASSERT_EQ(listenIP, "0.0.0.0");
     ASSERT_EQ(listenPort, "3333");
-    ASSERT_EQ(poolConfs.size(), 1);
+    ASSERT_EQ(poolConfs.size(), 1U);
     ASSERT_EQ(poolConfs[0].host_, "cn.ss.btc.com");
-    ASSERT_EQ(poolConfs[0].port_, 1800);
+    ASSERT_EQ(poolConfs[0].port_, 1800U);
     // ASSERT_EQ(poolConfs[0].upPoolUserName_, "kevin");
   }
 
@@ -76,14 +76,14 @@ TEST(Utils, Strings_parseConfJson) {
 
     ASSERT_EQ(listenIP, "127.0.0.1");
     ASSERT_EQ(listenPort, "1800");
-    ASSERT_EQ(poolConfs.size(), 2);
+    ASSERT_EQ(poolConfs.size(), 2U);
 
     ASSERT_EQ(poolConfs[0].host_, "cn.ss.btc.com");
-    ASSERT_EQ(poolConfs[0].port_, 1800);
+    ASSERT_EQ(poolConfs[0].port_, 1800U);
     // ASSERT_EQ(poolConfs[0].upPoolUserName_, "kevin");
 
     ASSERT_EQ(poolConfs[1].host_, "us.ss.btc.com");
-    ASSERT_EQ(poolConfs[1].port_, 3333);
+    ASSERT_EQ(poolConfs[1].port_, 3333U);
     // ASSERT_EQ(poolConfs[1].upPoolUserName_, "kevinus");
   }
 }
